refactor: refatorando controlers e services

+++ src/main/java/com/javanfood/javanfood/domain/service/CadastroCidadeService.java
@@ -1,47 +1,54 @@
 package com.javanfood.javanfood.domain.service;
 
-import com.javanfood.javanfood.domain.repository.CidadeRepository;
-import com.javanfood.javanfood.domain.repository.EstadoRespository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.stereotype.Service;
+
+import com.javanfood.javanfood.domain.exeption.EntidadeEmUsoExeption;
 import com.javanfood.javanfood.domain.exeption.EntidadeNaoEncontradaExeption;
 import com.javanfood.javanfood.domain.model.Cidade;
 import com.javanfood.javanfood.domain.model.Estado;
-import com.javanfood.javanfood.domain.model.Restaurante;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.stereotype.Service;
+import com.javanfood.javanfood.domain.repository.CidadeRepository;
+import com.javanfood.javanfood.domain.repository.EstadoRespository;
 
 @Service
 public class CadastroCidadeService {
-    @Autowired
-    CidadeRepository cidadeRepository;
+	private static final String MSG_CIDADE_NAO_ENCONTRADA = "Não existe cadastro de Cidade com esse código: %d";
 
-    @Autowired
-    EstadoRespository estadoRespository;
+	private static final String MSG_CIDADE_EM_USO = "Cidade de código: %d não pode ser removida, pois está em uso";
 
-    public Cidade salvar(Cidade cidade) {
-        Long estadoId = cidade.getEstado().getId();
-        Estado estado = estadoRespository.findById(estadoId)
-                .orElseThrow(() -> new EntidadeNaoEncontradaExeption(
-                        "Não existe cadastro de Estado com esse código: %d".formatted(estadoId))
-                );
+	@Autowired
+	CidadeRepository cidadeRepository;
 
-        cidade.setEstado(estado);
-        return cidadeRepository.save(cidade);
-    }
+	@Autowired
+	EstadoRespository estadoRespository;
 
-    public void excluir(Long cidadeId) {
-        try {
-            cidadeRepository.deleteById(cidadeId);
-        } catch (EmptyResultDataAccessException e) {
-            throw new EntidadeNaoEncontradaExeption("Não existe cadastro de Cidade com esse código: " + cidadeId);
-        } catch (DataIntegrityViolationException e) {
-            throw new EntidadeNaoEncontradaExeption("Cidade de código: " + cidadeId + " não pode ser removida. Está em uso.");
-        }
-    }
-    
-   
+	public Cidade buscaOuFalha(Long cidade_id) {
+		return cidadeRepository.findById(cidade_id)
+				.orElseThrow(() -> new EntidadeNaoEncontradaExeption(String.format(MSG_CIDADE_NAO_ENCONTRADA, cidade_id)));
+	}
+
+	public Cidade salvar(Cidade cidade) {
+		Long estadoId = cidade.getEstado().getId();
+		Estado estado = estadoRespository.findById(estadoId)
+				.orElseThrow(() -> new EntidadeNaoEncontradaExeption(
+						MSG_CIDADE_NAO_ENCONTRADA.formatted(estadoId)));
+
+		cidade.setEstado(estado);
+		return cidadeRepository.save(cidade);
+	}
+
+	public void excluir(Long cidadeId) {
+
+		if (!cidadeRepository.existsById(cidadeId)) {
+			throw new EntidadeNaoEncontradaExeption(String.format(MSG_CIDADE_NAO_ENCONTRADA, cidadeId));
+		}
+
+		try {
+			cidadeRepository.deleteById(cidadeId);
+			cidadeRepository.flush();
+		} catch (DataIntegrityViolationException e) {
+			throw new EntidadeEmUsoExeption(String.format(MSG_CIDADE_EM_USO, cidadeId));
+		}
+	}
 }
